refactor(edgeless): switch to gfx/viewport
