+++ packages/blocks/src/root-block/edgeless/utils/panning-utils.ts
@@ -1,7 +1,6 @@
 import type { PointerEventState } from '@blocksuite/block-std';
+import type { Viewport } from '@blocksuite/block-std/gfx';
 import type { IVec } from '@blocksuite/global/utils';
-
-import type { Viewport } from './viewport.js';
 
 const PANNING_DISTANCE = 30;
