+++ packages/blocks/src/root-block/edgeless/utils/zoom.ts
@@ -1,0 +1,5 @@
+export type ZoomAction = 'fit' | 'out' | 'reset' | 'in';
+export const ZOOM_MAX = 6.0;
+export const ZOOM_MIN = 0.1;
+export const ZOOM_STEP = 0.25;
+export const ZOOM_INITIAL = 1.0;
