+++ packages/framework/block-std/src/gfx/viewport.ts
@@ -13,11 +13,8 @@ function cutoff(value: number, ref: number, sign: number) {
   return value;
 }
 
-export type ZoomAction = 'fit' | 'out' | 'reset' | 'in';
 export const ZOOM_MAX = 6.0;
 export const ZOOM_MIN = 0.1;
-export const ZOOM_STEP = 0.25;
-export const ZOOM_INITIAL = 1.0;
 
 export class Viewport {
   protected _center: IPoint = { x: 0, y: 0 };
